CPAN-ready DHT commit (Chapter II) + [fix] Replaced carp calls with proper log callbacks + [fix] HAVEALL and HAVENONE packets were sending the wrong packet length...         - Which didn't matter seeing as FastPeers was disabled until now + [etc] Changed default conns_per_client: 300 => 200 + [etc] Changed N::B::new() arguments:       Old                           New       ------------------------------------------------------------------------       maximum_peers_per_client      conns_per_client       maximum_peers_per_session     conns_per_session       kbps_down                     max_dl_rate       kbps_up                       max_ul_rate       maximum_peers_half_open       max_halfopen       maximum_buffer_size           max_buffer_per_conn       maximum_requests_size         ul_slot_size       maximum_requests_per_peer     ul_slots_per_session
+ [etc] Changed accessors (incomplete list):
     Old Mutator                Getter                   Setter
     ---------------------------------------------------------------------------
    N::B
     maximum_buffer_size       get_max_buffer_per_conn set_max_buffer_per_conn
     kbps_down                 get_max_dl_rate          set_max_dl_rate
     kbps_up                   get_max_ul_rate          set_max_ul_rate
     maximum_peers_per_client  get_conns_per_client     set_conns_per_client
     maximum_peers_per_session get_conns_per_session    set_conns_per_session
     maximum_peers_half_open   get_max_halfopen         set_max_halfopen
     debug_level               get_debug_level          set_debug_level
     maximum_requests_per_peer get_ul_slots_per_session set_ul_slots_per_session
     maximum_requests_size     get_ul_slot_size         get_ul_slot_size
     peer_id                   get_peer_id              --NA--
     sockport                  get_sockport             --NA--
     sockaddr                  get_sockaddr             --NA--
     sessions                  get_sessions             --NA--

    N::B::S::Piece
     check                     get_cached_integrity
     verify                    get_verified_integrity

+ [etc] Removed accessors:
        - N::B->timeout( [NEWVAL] )
          This was the select timeout used in N::B::do_one_loop().  It has
          been removed completely in favor of an optional parameter passed to
          N::B::do_one_loop() itself.
+ [fix] t/data/torrents/miniswarm.torrent was badly broken. Now hand made
+ [new] t/data/etc/http_miniswarm.pl now takes a third param: timeout
+ [fix] t/500_full_swarm/* are now stable enough to be standard tests?
+ [etc] Several swarm tests have been removed.
        - Why was I going out of my way to be redundant?
+ [fix] DHT deals with socket open() failures a little better (UDP is fickle.)
+ [new] "Smart" Developer/Unstable build numbers and peer_id generation
+ [fix] N::B::S::Peer would sometimes (randomly) 'forget' to request the last
          piece (zero based problem)
+ [etc] Module::Build's create_makefile_pl set to traditional (disabled)
+ [etc] README updated to new API and added link to homepage in Bugs
+ [fix] Documentation has been brought up to date with current API and all
         internal and external links have been validated.
+ [fix] N::B::DHT workaround for http://groups.google.com/group/perl.cpan.testers/msg/bab8a8d99f2c9c2f
+ [new] N::B::S::Tracker::add_url() and N::B::S::Tracker::remove_url()
+ [etc] CPAN!

+++ t/data/etc/http_miniswarm.pl
@@ -4,89 +4,135 @@
 use strict;
 use warnings;
 use Socket;
-my ($seeds, $leeches) = @ARGV;
+my ($seeds, $leeches, $timeout) = @ARGV;
+$seeds   ||= 1;
+$leeches ||= 1;
+$timeout ||= 120;
+my $sprintf = q[%0] . length($leeches > $seeds ? $leeches : $seeds) . q[d];
 socketpair(my ($CHILD), my ($PARENT), AF_UNIX, SOCK_STREAM, PF_UNSPEC)
     or die "socketpair: $!";
+
 if (my $pid = fork) {
     my $port = undef;
     close $PARENT;
     chomp(my $line = readline $CHILD);
     if ($line =~ m[p:(\d+)]) { $port = $1 }
-    else                     { die; }    # ...um, I mean skip.
+    else                     { die; }    # XXX - ...um, I mean skip.
     close $CHILD;
-    {
+SKIP: {
         use Test::More;
         use File::Temp qw[];
         use lib q[../../lib];
+        use lib q[../../../lib];
         use Net::BitTorrent;
         use Net::BitTorrent::Util qw[compact];
-        plan tests => $seeds + $leeches;
+        plan tests => int($seeds + $leeches);
         my %client;
         my $test_builder = Test::More->builder;
 
         for my $chr (1 .. $seeds) {
-            $client{q[seed_] . $chr} = new Net::BitTorrent(
-                                 {LocalAddr => q[127.0.0.1], Timeout => 0.1});
-            skip(sprintf(q[Failed to create seed_%s], $chr),
-                 $test_builder->{Expected_Tests} - $test_builder->{Curr_Test}
+            $chr = sprintf $sprintf, $chr;
+            $client{q[seed_] . $chr}
+                = new Net::BitTorrent({LocalAddr => q[127.0.0.1]});
+            skip_all(sprintf(q[Failed to create seed_%s], $chr),
+                     $test_builder->{q[Expected_Tests]}
+                         - $test_builder->{q[Curr_Test]}
             ) if not $client{q[seed_] . $chr};
+
+         #{    # DEBUG
+         #$client{q[seed_] . $chr}->set_callback(q[peer_connect],
+         #                           sub { shift; warn shift; });
+         #$client{q[seed_] . $chr}->set_callback(q[peer_disconnect],
+         #                           sub { shift; shift; warn shift; });
+         #$client{q[seed_] . $chr}->set_callback(q[tracker_error],
+         #                           sub { shift; shift; warn shift; });
+         #$client{q[seed_] .$chr}->set_callback(q[peer_incoming_packet], sub {
+         #    shift; shift;
+         #    use Data::Dump qw[pp];
+         #warn q[Packet!] . pp shift});
+         #$client{q[seed_] . $chr}
+         #    ->set_callback(q[log], sub { shift; shift; warn shift; });
+         #$client{q[seed_] . $chr}->set_debug_level(1000);
+         #}
             my $session = $client{q[seed_] . $chr}->add_session(
                            {path => q[./t/data/torrents/miniswarm.torrent],
                             base_dir => q[./t/data/miniswarm/]
                            }
             );
             skip(sprintf(q[Failed to load session for seed_%s], $chr),
-                 $test_builder->{Expected_Tests} - $test_builder->{Curr_Test}
+                 $test_builder->{q[Expected_Tests]}
+                     - $test_builder->{q[Curr_Test]}
             ) if not $session;
-            ok(scalar($session->complete), sprintf(q[seed_%s ok], $chr));
+            ok(scalar($session->get_complete), sprintf(q[seed_%s ok], $chr));
+            skip(sprintf(q[Failed to load session for seed_%s], $chr),
+                 $test_builder->{q[Expected_Tests]}
+                     - $test_builder->{q[Curr_Test]}
+            ) if not $session->get_complete;
             my $tracker = qq[http://127.0.0.1:$port/announce];
             $session->add_tracker([$tracker]);
+            $client{q[seed_] . $chr}->do_one_loop(0.1);    # let them announce
         }
         for my $chr (1 .. $leeches) {
-            $client{$chr} = new Net::BitTorrent(
-                                 {LocalAddr => q[127.0.0.1], Timeout => 0.1});
-            $client{$chr}->set_callback(q[peer_disconnect],
-                                        sub { shift; shift; warn shift; });
-
-            #$client{$chr}->set_callback(q[log],
-            #                       sub { shift; shift; warn shift; });
-            #$client{$chr}->debug_level(1000);
+            $chr = sprintf $sprintf, $chr;
+            $client{$chr} = new Net::BitTorrent({LocalAddr => q[127.0.0.1]});
             skip(sprintf(q[Failed to create leech_%s], $chr),
-                 $test_builder->{Expected_Tests} - $test_builder->{Curr_Test}
+                 $test_builder->{q[Expected_Tests]}
+                     - $test_builder->{q[Curr_Test]}
             ) if not $client{$chr};
+
+#{    # DEBUG
+#$client{$chr}->set_callback(q[peer_connect],
+#                           sub { shift;  warn shift; });
+#$client{$chr}->set_callback(q[peer_disconnect],
+#                           sub { shift; shift; warn shift; });
+#$client{$chr}->set_callback(q[tracker_error],
+#                           sub { shift; shift; warn shift; });
+#$client{$chr}->set_callback(q[peer_outgoing_request], sub {warn q[Request!]});
+#$client{$chr}->set_callback(q[peer_incoming_packet], sub {
+#    shift; shift;
+#    use Data::Dump qw[pp];
+#    warn q[Packet!] . pp shift
+#    });
+#$client{$chr}
+#    ->set_callback(q[log], sub { shift; shift; warn shift; });
+#$client{$chr}->set_debug_level(1000);
+#}
             $client{$chr}->set_callback(
                 q[piece_hash_pass],
                 sub {
                     my ($self, $piece) = @_;
-                    my $session = $piece->session;
+                    my $session = $piece->get_session;
                     my $completion = (
                         (((scalar grep {
-                               $_->priority and $_->check
-                               } @{$session->pieces}
-                          ) / (scalar @{$session->pieces})
+                                       $_->get_priority
+                                   and $_->get_cached_integrity
+                               } @{$session->get_pieces}
+                          ) / (scalar @{$session->get_pieces})
                          )
                         ) * 100
                     );
-                    my $line =
-                        sprintf(q[(%02d|%02d) [%s] %.2f%%],
-                                $chr,
-                                $piece->index,
-                                join(q[],
-                                     map ((   $_->check
-                                            ? $piece->index == $_->index
-                                                    ? q[*]
-                                                    : q[|]
-                                            : scalar $_->working ? q[.]
-                                            : q[ ]
-                                          ),
-                                          @{$session->pieces})
-                                ),
-                                $completion
-                        );
-
-                    #diag($line);
-                    ok($session->complete, sprintf(q[peer_%s complete], $chr))
-                        if $session->complete;
+                    diag(sprintf(
+                               q[(%02d|%02d) [%s] %.2f%%],
+                               $chr,
+                               $piece->get_index,
+                               join(
+                                   q[],
+                                   map ((  $_->get_cached_integrity
+                                         ? $piece->get_index == $_->get_index
+                                                 ? q[*]
+                                                 : q[|]
+                                         : scalar $_->get_working ? q[.]
+                                         : q[ ]
+                                       ),
+                                       @{$session->get_pieces})
+                               ),
+                               $completion
+                         )
+                        )
+                        if $ENV{q[RELEASE_TESTING]};
+                    ok($session->get_complete,
+                        sprintf(q[peer_%s complete], $chr))
+                        if $session->get_complete;
                     return;
                 }
             );
@@ -103,19 +149,29 @@ if (my $pid = fork) {
                            }
                 );
             skip(sprintf(q[Failed to load session for leech_%s], $chr),
-                 $test_builder->{Expected_Tests} - $test_builder->{Curr_Test}
+                 $test_builder->{q[Expected_Tests]}
+                     - $test_builder->{q[Curr_Test]}
             ) if not $session;
             my $tracker = qq[http://127.0.0.1:$port/announce];
             $session->add_tracker([$tracker]);
         }
-        while ($test_builder->{Curr_Test} < $test_builder->{Expected_Tests}) {
-            grep { $_->do_one_loop } values %client;
+        while ($test_builder->{q[Curr_Test]}
+               < $test_builder->{q[Expected_Tests]})
+        {   grep { $_->do_one_loop(0.1); } values %client;
+            skip(q[This is taking too long and I have a train to catch.],
+                 (      $test_builder->{q[Expected_Tests]}
+                      - $test_builder->{q[Curr_Test]}
+                 )
+            ) if (int(time - $^T) > $timeout);
         }
-        grep {
-            $_->remove_session($_->sessions->[0])
-                if scalar @{$_->sessions}
-        } values %client;
-        kill 9, $pid;
+
+        END {
+            kill 9, $pid if defined $pid;
+            grep {
+                $_->remove_session($_->get_sessions->[0])
+                    if scalar @{$_->get_sessions}
+            } values %client;
+        }
         exit;
     }
 }
